Installation of a library to use AWS s3
We will eventually want to save backups on s3 instead of Dropbox. The
"s3" disk is already defined in config/filesystems.php, but it requires
a special library to be able to use it.

Install the league/flysystem-aws-s3-v3 library and update .env.example
to add the AWS global variables.